+++ source/refined-github.ts
@@ -220,3 +220,4 @@ import './features/linkify-text.js';
 import './features/actions-run-removal.js';
 import './features/new-milestone-button.js';
 import './features/no-self-reference.js';
+import './features/mobile-conversation-headers.js';
